refactor: Atualiza PatientController para utilizar Redis nas consultas

+++ app/Http/Controllers/PatientController.php
@@ -2,15 +2,12 @@
 
 namespace App\Http\Controllers;
 
-use App\Http\Resources\PatientCollection;
 use App\Http\Resources\PatientResource;
 use App\Models\Address;
 use App\Models\Patient;
-use Illuminate\Foundation\Application;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
-use Illuminate\Support\Facades\App;
-use Illuminate\Support\Facades\Config;
+use Illuminate\Support\Facades\Cache;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Response;
 
@@ -37,8 +34,8 @@ class PatientController extends Controller
     /**
      * Store a newly created resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
+     * @param Request $request
+     * @return JsonResponse
      */
     public function store(Request $request)
     {
@@ -46,24 +43,13 @@ class PatientController extends Controller
         DB::beginTransaction();
         try{
             $patient = new Patient();
-            $patient->full_name = $request->full_name;
-            $patient->mother_name = $request->mother_name;
-            $patient->date_of_birth = $request->date_of_birth;
-            $patient->cpf = $request->cpf;
-            $patient->cns = $request->cns;
+            $patient->fill($request->all());
+            $patient->save();
 
             $address = new Address();
-            $address->cep = $request->address['cep'];
-            $address->address = $request->address['address'];
-            $address->number = $request->address['number'];
-            $address->complement = $request->address['complement'];
-            $address->neighborhood = $request->address['neighborhood'];
-            $address->city = $request->address['city'];
-            $address->state = $request->address['state'];
-
+            $address->fill($request->address);
+            $address->patient_id = $patient->id;
             $address->save();
-            $patient->address_id = $address->id;
-            $patient->save();
 
             DB::commit();
             return response()->json(['message' => 'Paciente criado com sucesso', 'data' => new PatientResource($patient)], 201);
@@ -78,15 +64,13 @@ class PatientController extends Controller
      * Display the specified resource.
      *
      * @param  int  $id
-     * @return \Illuminate\Http\Response
+     * @return JsonResponse
      */
     public function show($id)
     {
-        $patient = Patient::find($id);
-
-        if (!$patient) {
-            return response()->json(['error' => 'Paciente não encontrado'], 404);
-        }
+        $patient = Cache::remember('patient_'.$id, 3600, function () use ($id) {
+            return Patient::findOrFail($id);
+        });
 
         return response()->json(new PatientResource($patient));
     }
@@ -94,38 +78,37 @@ class PatientController extends Controller
     /**
      * Update the specified resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
+     * @param Request $request
      * @param  int  $id
-     * @return \Illuminate\Http\Response
+     * @return JsonResponse
      */
     public function update(Request $request, $id)
     {
         DB::beginTransaction();
         try{
             $patient = Patient::find($id);
-            $patient->full_name = $request->full_name;
-            $patient->mother_name = $request->mother_name;
-            $patient->date_of_birth = $request->date_of_birth;
-            $patient->cpf = $request->cpf;
-            $patient->cns = $request->cns;
+
+            if (!$patient) {
+                return response()->json(['error' => 'Paciente não encontrado'], 404);
+            }
+
+            $patient->fill($request->all());
             $patient->save();
 
-            if(isset($request->address)){
-                $address = Address::find($patient->address_id);
-                $address->cep = $request->address['cep'];
-                $address->address = $request->address['address'];
-                $address->number = $request->address['number'];
-                $address->complement = $request->address['complement'];
-                $address->neighborhood = $request->address['neighborhood'];
-                $address->city = $request->address['city'];
-                $address->state = $request->address['state'];
-                $address->save();
+            if (isset($request->address)) {
+                $address = Address::updateOrCreate(
+                    ['patient_id' => $patient->id],
+                    $request->address
+                );
             }
+
+            // Atualiza o cache
+            Cache::put('patient_' . $patient->id, $patient, 3600);
 
             DB::commit();
             return response()->json(['message' => 'Paciente atualizado com sucesso', 'data' => new PatientResource($patient)], 201);
 
-        } catch (Throwable $e) {
+        } catch (\Exception $e) {
             DB::rollback();
             return response()->json(['message' => 'Erro ao atualizar paciente: ' . $e->getMessage()], 500);
         }
@@ -135,11 +118,17 @@ class PatientController extends Controller
      * Remove the specified resource from storage.
      *
      * @param  int  $id
-     * @return \Illuminate\Http\Response
+     * @return void
      */
     public function destroy($id)
     {
         $patient = Patient::find($id);
-        $patient->delete();
+
+        if($patient){
+            $patient->delete();
+        }
+
+        // Remove o paciente do cache
+        Cache::forget('patient_' . $id);
     }
 }
